requirements.txt
Flask==2.2.2
pyjwt
